LPS-146610 Restructure assignments components
