+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-designer-web/src/main/resources/META-INF/resources/designer/js/definition-builder/diagram-builder/components/sidebar/sections/assignments/Assignments.js
@@ -9,39 +9,69 @@
  * distribution rights of the Software.
  */
 
-import ClayButton from '@clayui/button';
-import PropTypes from 'prop-types';
-import React, {useContext} from 'react';
+import React, {useContext, useEffect, useState} from 'react';
 
 import {DiagramBuilderContext} from '../../../../DiagramBuilderContext';
-import SidebarPanel from '../../SidebarPanel';
-import CurrentAssignments from './CurrentAssignments';
+import AssetCreator from './select-assignment/AssetCreator';
+import ResourceActions from './select-assignment/ResourceActions';
+import Role from './select-assignment/Role';
+import RoleType from './select-assignment/RoleType';
+import ScriptedAssignment from './select-assignment/ScriptedAssignment';
+import {SelectAssignment} from './select-assignment/SelectAssignment';
+import User from './select-assignment/User';
+import {getAssignmentType} from './utils';
 
-const Assignments = ({setContentName}) => {
-	const {selectedItem} = useContext(DiagramBuilderContext);
-
-	return (
-		<SidebarPanel panelTitle={Liferay.Language.get('assignments')}>
-			{!selectedItem?.data?.assignments ? (
-				<ClayButton
-					className="mr-3"
-					displayType="secondary"
-					onClick={() => setContentName('assignments')}
-				>
-					{Liferay.Language.get('new')}
-				</ClayButton>
-			) : (
-				<CurrentAssignments
-					assignments={selectedItem.data.assignments}
-					setContentName={setContentName}
-				/>
-			)}
-		</SidebarPanel>
-	);
+const assignmentSectionComponents = {
+	assetCreator: AssetCreator,
+	resourceActions: ResourceActions,
+	roleId: Role,
+	roleType: RoleType,
+	scriptedAssignment: ScriptedAssignment,
+	user: User,
 };
 
-Assignments.propTypes = {
-	setContentName: PropTypes.func.isRequired,
+const Assignments = (props) => {
+	const {selectedItem} = useContext(DiagramBuilderContext);
+	const assignments = selectedItem?.data?.assignments;
+
+	const assignmentType = getAssignmentType(assignments);
+
+	const [section, setSection] = useState(assignmentType || 'assetCreator');
+	const [sections, setSections] = useState([{identifier: `${Date.now()}-0`}]);
+
+	const AssignmentSectionComponent = assignmentSectionComponents[section];
+
+	useEffect(() => {
+		if (assignmentType === 'user') {
+			setSections(assignments.sectionsData);
+		}
+		// eslint-disable-next-line react-hooks/exhaustive-deps
+	}, []);
+
+	return (
+		<>
+			<SelectAssignment
+				section={section}
+				setSection={setSection}
+				setSections={setSections}
+			/>
+
+			{sections.map(({identifier}, index) => {
+				return (
+					AssignmentSectionComponent && (
+						<AssignmentSectionComponent
+							{...props}
+							identifier={identifier}
+							index={index}
+							key={`section-${identifier}`}
+							sectionsLength={sections?.length}
+							setSections={setSections}
+						/>
+					)
+				);
+			})}
+		</>
+	);
 };
 
 export default Assignments;
