+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-designer-web/src/main/resources/META-INF/resources/designer/js/definition-builder/diagram-builder/components/sidebar/sections/assignments/CurrentAssignments.js
@@ -16,7 +16,7 @@ import React, {useContext, useEffect, useState} from 'react';
 
 import lang from '../../../../../util/lang';
 import {DiagramBuilderContext} from '../../../../DiagramBuilderContext';
-import {options} from './SelectAssignment';
+import {options} from './select-assignment/SelectAssignment';
 import {getAssignmentType} from './utils';
 
 const CurrentAssignments = ({assignments, setContentName}) => {
